Apply style fixes for modus-wc-button
