+++ src/components/modus-wc-button/modus-wc-button.stories.ts
@@ -2,6 +2,7 @@ import { withActions } from '@storybook/addon-actions/decorator';
 import { Meta, StoryObj } from '@storybook/web-components';
 import { html } from 'lit';
 import { ifDefined } from 'lit/directives/if-defined.js';
+import { createShadowTestHost } from '../../stories/utilities/shadow-test-host';
 import { DaisySize } from '../types';
 
 interface ButtonArgs {
@@ -157,99 +158,14 @@ export const IconLeftAndRightButton: Story = {
   },
 };
 
-if (!customElements.get('modus-shadow-test-host')) {
-  console.log('Defining modus-shadow-test-host');
-  class ModusShadowTestHost extends HTMLElement {
-    private observer!: MutationObserver;
-    static get observedAttributes() {
-      return ['tag', 'props', 'attrs', 'innerhtml'];
-    }
-
-    private shadow: ShadowRoot;
-    private child: HTMLElement | null = null;
-
-    constructor() {
-      super();
-      this.shadow = this.attachShadow({ mode: 'open' });
-    }
-
-    connectedCallback() {
-      this.copyThemeAttributes();
-      this.observer = new MutationObserver(() => this.copyThemeAttributes());
-      this.observer.observe(document.documentElement, {
-        attributes: true,
-        attributeFilter: ['data-theme', 'data-mode'],
-      });
-      this.renderChild();
-    }
-
-    copyThemeAttributes() {
-      const html = document.documentElement;
-      const theme = html.getAttribute('data-theme');
-      const mode = html.getAttribute('data-mode');
-      if (theme !== null) {
-        this.setAttribute('data-theme', theme);
-      } else {
-        this.removeAttribute('data-theme');
-      }
-      if (mode !== null) {
-        this.setAttribute('data-mode', mode);
-      } else {
-        this.removeAttribute('data-mode');
-      }
-    }
-
-    attributeChangedCallback() {
-      this.renderChild();
-    }
-
-    renderChild() {
-      console.log('Rendering child in shadow DOM');
-      // Remove previous child
-      if (this.child) this.shadow.removeChild(this.child);
-
-      const tag = this.getAttribute('tag') || 'div';
-      const props = this.getAttribute('props')
-        ? JSON.parse(this.getAttribute('props')!)
-        : {};
-      const attrs = this.getAttribute('attrs')
-        ? JSON.parse(this.getAttribute('attrs')!)
-        : {};
-      const innerHTML = this.getAttribute('innerhtml') || '';
-
-      const el = document.createElement(tag);
-
-      // Set attributes
-      Object.entries(attrs).forEach(([k, v]) =>
-        el.setAttribute(k, v as string)
-      );
-      // Set properties
-      Object.entries(props).forEach(([k, v]) => ((el as any)[k] = v));
-
-      el.innerHTML = innerHTML;
-      this.shadow.appendChild(el);
-      this.child = el;
-    }
-  }
-  console.log('Defining modus-shadow-test-host');
-  customElements.define('modus-shadow-test-host', ModusShadowTestHost);
-} else {
-  console.log('modus-shadow-test-host already defined');
-}
+createShadowTestHost();
 
 export const RenderedInShadowDom: Story = {
   render: () => {
-    // Example: Render modus-wc-button with attributes and innerHTML
     const attrs = JSON.stringify({
       color: 'primary',
     });
     const innerhtml = 'Click me';
-    console.log(
-      'Rendering in shadow DOM with attrs:',
-      attrs,
-      'and innerHTML:',
-      innerhtml
-    );
     return html`
       <modus-shadow-test-host
         tag="modus-wc-button"
